refactor: move label functions to helicity.naming
